Resolve Schema#requires? returning incorrect results for references
This resolves https://github.com/kevindew/openapi3_parser/issues/20

This code relied on comparing node equality to determine whether a
schema property was required or not. This was flawed as referenced
schemas are considered equal and meant that a property that is not
required could return true from #requires?

To resolve this I've switched the code to consider the node context.
This will consider the node's placement in the document when deciding
whether a schema is required or not.

This scenario seems to illustrate something of a design smell as it
could be confusing to have code such as:

```
property_a = schema.properties["a"]
property_b = schema.properties["b"]

schema.requires?(property_a) # => true
schema.requires?(property_b) # => false
property_a == property_b # => true
```

However this seems to be the best way out of this particular problem,
where the method is clearly returning the wrong result. If this method
produces further problems it may be best to only accept the string of
the property name rather than trying to consider the schema directly.